# XWIKI-22979: Access to pages is slow under high load on the security cache