--- a/xwiki-platform-core/xwiki-platform-flamingo/xwiki-platform-flamingo-skin/xwiki-platform-flamingo-skin-test/xwiki-platform-flamingo-skin-test-docker/src/main/resources/META-INF/components.txt
+++ b/xwiki-platform-core/xwiki-platform-flamingo/xwiki-platform-flamingo-skin/xwiki-platform-flamingo-skin-test/xwiki-platform-flamingo-skin-test-docker/src/main/resources/META-INF/components.txt
@@ -1,3 +1,4 @@
 org.xwiki.test.CustomUserUpdatedDocumentEventListener
 org.xwiki.test.TestMacro
 org.xwiki.test.SleepScriptService
+org.xwiki.test.SecurityCachePerformanceTestScriptService