--- a/xwiki-platform-core/xwiki-platform-flamingo/xwiki-platform-flamingo-skin/xwiki-platform-flamingo-skin-test/xwiki-platform-flamingo-skin-test-docker/src/test/it/org/xwiki/flamingo/test/docker/AllIT.java
+++ b/xwiki-platform-core/xwiki-platform-flamingo/xwiki-platform-flamingo-skin/xwiki-platform-flamingo-skin-test/xwiki-platform-flamingo-skin-test-docker/src/test/it/org/xwiki/flamingo/test/docker/AllIT.java
@@ -218,4 +218,10 @@ public class AllIT
     class NestedPageReadyIT extends PageReadyIT
     {
     }
+
+    @Nested
+    @DisplayName("Security Cache Stress Tests")
+    class NestedSecurityCacheStressIT extends SecurityCacheStressIT
+    {
+    }
 }