--- a/xwiki-platform-core/xwiki-platform-security/xwiki-platform-security-authorization/xwiki-platform-security-authorization-api/src/main/java/org/xwiki/security/authorization/cache/internal/DefaultSecurityCache.java
+++ b/xwiki-platform-core/xwiki-platform-security/xwiki-platform-security-authorization/xwiki-platform-security-authorization-api/src/main/java/org/xwiki/security/authorization/cache/internal/DefaultSecurityCache.java
@@ -104,7 +104,7 @@ public class DefaultSecurityCache implements SecurityCache, Initializable
     private final AtomicBoolean invalidationInProgress = new AtomicBoolean();
 
     /** Fair write lock. */
-    private final Lock writeLock = new ReentrantLock(true);
+    private final Lock writeLock = new ReentrantLock();
 
     private final ReadWriteLock invalidationReadWriteLock = new ReentrantReadWriteLock(true);