# Jira URL

<!-- Add the link to the corresponding JIRA issue referenced in a commit message. Unless this is a [Misc] commit,
see https://dev.xwiki.org/xwiki/bin/view/Community/DevelopmentPractices#HRule:Don27tcreateunnecessaryissues
-->

https://jira.xwiki.org/browse/XWIKI-22979

# Changes

## Description

<!-- Describe the main changes brought in this PR. -->

* Remove the read lock of the security cache to avoid blocking read access during cache writes.
* Switch from a ReadWriteLock to a simple write lock as ReadWriteLocks might be slow, and we don't really need it anymore.
* Remove the useless check of the internal entries for security access entries to avoid needlessly acquiring the write lock.
* Improve the concurrency behavior of upgrading entries to user entries to ensure that removing the read lock is safe.
* Replace the invalidation lock used to prevent loading outdated data by an invalidation counter. This avoids holding any locks in the security cache while a document is potentially loaded from the database. Before that, it could happen that while a remove operation is waiting on the lock, a database load is executed under the invalidation read lock. That would block both the invalidation and all further cache loads as the invalidation read lock cannot be obtained any more as soon as a thread is waiting for the invalidation write lock.
* Update the security cache loader and the affected tests.
* Make the write lock in the security cache unfair for increased performance.
* Add a stress test for the security cache.


## Clarifications

<!-- Provide extra hints to make it easier to understand the PR. Those could be:
* Explanation of choices made in this PR
* Anchor towards extra resources needed to understand the context of this PR (e.g., a forum proposal).
* Links to other issues this issue depends on
-->

The discarding of writes to the security cache could cause performance degradation in certain cases where a lot of documents are updated, in particular if the loaded document isn't kept in the document cache. However, this seems too unlikely that it would warrant a much more complex security cache (it is already complex enough). This PR implements approach 1 in [the design page](https://design.xwiki.org/xwiki/bin/view/Proposal/SecurityCacheRedesign#HInvalidationLock), to reduce the number of discarded additions, we would need to implement approach 2.

The stress test first creates 8k pages in a hierarchy of 3 levels (`A.B.C.WebHome`) and 20 users, and then executes 12 extra threads. They perform the following access control checks:

1. Thread 1-10 check access right for guest on the first 1k pages in an endless loop.
2. Thread 11 checks access right for guest on all pages in random order in an endless loop.
3. Thread 12 checks access right for all 20 users on all pages in another random order in an endless loop.

The main thread saves a random subset of pages while these loops are running. Then it sends a stop signal to the extra threads and builds the average times of these access control checks. To warm up the caches, the benchmark runs twice: first saving 100 pages, then saving 500 pages. As the security cache has only a capacity of 10k items and the document cache a capacity of 500 items, there are quite some cache misses and loads. 

I've executed the stress test both with the old and the new code on my laptop. Here are the results:

| | With old code and MariaDB 11.5: | | With new code and MariaDB 11.5: | |
| - | - | -  | - | - |
| | Test run 1 (100 pages):| Test run 2 (500 pages): | Test run 1 (100 pages):| Test run 2 (500 pages): |
| Top1kAverage: | 0.42ms   | 0.21ms  | 0.046ms  | 0.014ms  |
| RandomGuestAverage: | 2.82ms   |  1.13ms  | 5.59ms  | 4.74ms  |
| RandomAllUsersAverage: | 0.77ms | 0.90ms  |  0.43ms  | 0.35ms |
| Average saving time:  | 104.61ms | 95.18ms | 122.9ms | 103.68ms |

This table might need some explanation as some of these results seem counter-intuitive. First, as expected/desired, removing the read lock lead to quite a performance boost for cached reads. A factor 14 improvement seems quite good, in particular as the actual check in the security cache is only a part of that running time. I haven't recorded this information, but I can only assume that the worst-case running time before the changes was even worse.

Now regarding the other running times, in particular for guests, we see an almost factor 4 worse running time. I have two explanations for this:

1. The cache isn't fully deterministic, seeing much more hits for the top 1k pages might modify its behavior.
2. Due to the invalidation counter, we indeed don't insert some items into the cache. The guest test is more affected by this than the user cache due to a special behavior of the security cache: the access control check first checks if the page is in the security cache without security rules. If this is the case, it walks the hierarchy up until it finds an item with rules. As the benchmark doesn't generate any rights, this means that in the test with users, if the item for the page is inserted into the cache, all subsequent requests are most likely hits as the access control items on the root of the cache should never be evicted.

Overall, I have the impression that the new code works as expected from these tests.


# Screenshots & Video

<!-- If this PR introduces any UI change, it's recommended to highlight it with before/after screenshots 
or even a screen recording for complex interactions. 
-->

No UI changes.

# Executed Tests

<!-- Especially important for regression fixes. 
Indicate how changes were tested (e.g., what maven commands were run to validate them).
-->

```
mvn clean install -Pquality,integration-tests,docker -pl :xwiki-platform-flamingo-skin-test-docker,:xwiki-platform-security-authorization-api -Dit.test=SecurityCacheStressIT -Dxwiki.test.ui.database=mariadb -Dxwiki.test.ui.databaseTag=11.5
```

# Expected merging strategy

* Prefers squash: Yes <!-- No — Explain why. -->
* Backport on branches:
  * Difficult, I guess maybe after some time?